python-modules/requirements.txt
influxdb
paho-mqtt
python-dateutil
requests
-e libs/influx
-e libs/docker_secrets
